For the sake of showing the use of the MPL 2.0 OSMP sensor packaging in combination with a model implementation and looking at the weak copy left of MPL, we split the function into a different file
